Cancel the use of object pooling mechanism and merge some files.
